Added functionality to checker and FileCheck
FileCheck
Constructor takes a Path object and, through the method FileChecker,
compares the file represented by that path to the file represented
by the Path supplied as argument. The files are compared byte by byte
using a BufferedInputStream

checker
Creates a SimpleFileVisitor and overrides PreDirectoryVisit,
PostDirectoryVisit, and FileVisit. ArrayList paths is filled with each
visit to a new directory, the files are then compared against each other
using FileCheck, and the filenames are printed if they have a duplicate.
paths is then cleared. Note that currently each file is only compared with
the file immediately following. This is a proof of concept; ultimately
each file will be compared against every other file in the same
directory.
Ignores files that begin with '.', but that requires that only an
absolute pathname can be provided at the command line.